Add very basic unit test and rake task
The name of the test indicates that at the moment, I am only getting the parameters for the vCloud Launcher integration test. That - and the names - will change.

Also, the unit testing exposed that I needed to require yaml - this previously worked as I was testing against vCloud Launcher which included yaml. Unit testing FTW!