Use env var to determine org parameters to use
When running one of the vCloud Tools, unless you've exported the FOG_CREDENTIAL, you pass it in at the beginning of the command. This commit makes the tests work in the same way. For example if you want to run the tests against the test-organisation, you must have this set-up in your .fog file, and you then would run the test `FOG_CREDENTIAL=test-organisation bundle exec rake integration`.

The disadvantage of this would be that everyone would need to have the same name for organisations in their .fog file if they are going to share parameters, but I don't think that's a major drawback as it's easy enough to change.