Rename catalog_item as vapp_template
For some reason it has been called `catalog_item` in all our configuration code, but in vCloud Director it is called a vApp Template. This renames the testing code to use the correct name. There is an outstanding story to change the actual configuration code as well: https://www.pivotaltracker.com/story/show/70452106.