Change network config parameter names
For consistency, from `network1` to `network_1`.

Doing this I also noticed one of the testing config parameters was 1 instead of 2, but this is not tested, only the launcher config is fully tested so it hadn't come up before. Either use this config in a test or remove it, is a task for another day.