Dynamically generate param attribute readers
Dynamically generate attribute reader methods for the test parameters
read in from the YAML-formatted configuration file.

Implementing these methods dynamically means that we don't need to
manually add a new method and bump the version of this gem each time we
change test parameters in the integration tests.

Note that we're using `send` because `define_method` is a private method
and we're calling it from inside the class initalizer.