Throw friendly error message if param undefined
Throw a friendly error message using `method_missing` if an undefined
method is called on the TestParameters class, which usually indicates
that an integration test is trying to use a test parameter that has not
been defined or added to the YAML-formatted configuration file consumed
by vCloud Tools Tester.

Add new instance variable, `@config_file`, so that we can access the
name of the configuration file and use it in the error message in
`method_missing`.

* * *

Also amends the test to check for the new error message. Because we need
to match the error string, we use `let` to define `config_file` in the
tests so that it can be accessed both from inside the test and by the
other `let` declaration.