Confirm fixture networks available to vDCs
Add a new method
`Vcloud::Tools::Tester::FixtureParameters#network_available_to_correct_vdc?`
that checks if existing fixture networks are available to the vDCs used
by the tests.

This is necessary because the call we make to retrieve the network
attributes (line 38) contains no information on which vDCs the networks
are available to. Therefore, we make a second call to the Fog model
interface to query which vDCs are present in the environment and we
match the IDs of the networks available in each vDC to the networks we
found.

As part of this, `network_config_matches_expected` has been renamed
`network_config_matches_expected?` and now receives the network as an
object rather than a hash.

A new method, `network_matches_expected?`, has also been added. This
calls `network_config_matches_expected?` and
`network_available_to_correct_vdc?`.

Also creates an instance variable, `@vcloud_api`, so that the instance
of the model interface can be re-used.