[#74863398] Include `or` logic in case statement
Ruby allows for case statements to take multiple values in the `when`
statement. If any of these values is true, the script continues. If not, it
errors. Rather than using `or`, the syntax makes use of a comma to specify multiple
values of equal weighting in the `or`.

`,` is the equivalent of `||` in an `if` statement.