Broaden dependency on vCloud Core
Loosen the runtime dependency on vCloud Core so that we can use any
version of that gem greater or equal to version 0.5.0 and less than
version 1.0.

This is to break the current deadlock we have between versions due to
this circular dependency (vCloud Core is dependent on vCloud Tools
Tester). We can't release version 0.6.0 of vCloud Core because vCloud
Tools Tester is not yet released and it relies on vCloud Core 0.5.x.
Likewise, we can't release vCloud Tools Tester because it's dependent on
vCloud Core 0.6.