Bump version
This is required because of the circular dependency between vCloud Core and vCloud Tools Tester. In order to remove the temporary fog files that earlier versions of vCloud Tools Tester depended on I need to be able to have vCloud Core depend on a version of vCloud Tools Tester that doesn't require them.