Make network fixtures optional
In many integration tests, we do not need to have the network
fixtures present. This commit makes the requirement for having
a network fixture for (eg) network_1 or network_2 dependent
on the developer specifying "network_1" or "network_2" in the
expected_user_params array passed to TestSetup.

NB: This commit keeps the tests as-is for the 'both networks present'
    case, for simplicity. Futher tests and fixes for the 'no networks'
    case will be added in a subsequent commit.