Handle the case where no fixtures are needed
If there are no 'network_1' or 'network_2' parameters
specified in our expected_user_params input, then this
flags that we do not need to worry about the presence of
these fixtures for this test.

This creates the valid situation where there are no @fixtures and
no fixture_params, which previously was considered an error.